+++ src/promptflow/promptflow/_sdk/operations/_run_submitter.py
@@ -375,7 +375,8 @@ class RunSubmitter:
             result.update(
                 {
                     "run.outputs": self.run_operations._get_outputs_path(run.run),
-                    "run.inputs": self.run_operations._get_inputs_path(run.run),
+                    # to align with cloud behavior, run.inputs should refer to original data
+                    "run.inputs": self.run_operations._get_data_path(run.run),
                 }
             )
         return {k: str(Path(v).resolve()) for k, v in result.items() if v is not None}