+++ src/promptflow/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features Added
 
 - [SDK/CLI] Support `pfazure flow create`. Create a flow on Azure AI from local flow folder.
+- [SDK/CLI] Changed column mapping `${run.inputs.xx}`'s behavior, it will refer to run's data columns instead of run's inputs columns.
 
 ### Bugs Fixed